review: rr3-objects 修复 NeedToCreate Set 操作码错位/HINCRBY 溢出回绕/短值覆写 usize 下溢/随机取样 KOverN 双分支/SPOP 单枚随机化/负 COUNT 扫描语义/HEXPIREAT 溢出饱和/HRANDFIELD i32 口径

+++ node/wserver/src/objects/set/set_object.rs
@@ -190,7 +190,7 @@ impl SetObject {
   /// libs/server/Objects/Set/SetObject.cs:Scan
   ///
   /// 与 hash 不同：单条目形态，count 不翻倍；`cursor == Set.Count` 即归零
-  pub fn scan(&self, start: i64, count: usize, pattern: &[u8]) -> (Vec<Vec<u8>>, i64) {
+  pub fn scan(&self, start: i64, count: i64, pattern: &[u8]) -> (Vec<Vec<u8>>, i64) {
     let mut items: Vec<Vec<u8>> = Vec::new();
     let mut cursor = start;
 
@@ -212,7 +212,8 @@ impl SetObject {
 
       cursor += 1;
 
-      if items.len() == count {
+      // C# 以相等判断截断（负 COUNT 恒不命中 → 全量遍历，1:1 保留）
+      if items.len() as i64 == count {
         break;
       }
     }
